ecommerce_platform/ecommerce_platform/user_management/requirements.txt
Flask==1.1.4
Flask-RESTful==0.3.9
Werkzeug==1.0.1
psycopg2-binary==2.9.1







